Flip wiring of Dns.SYSTEM
Previously Dns.SYSTEM was a constant, and Platform.platformDns() was a
hook to return something different.

With this change, Dns.SYSTEM forwards to whatever Platform.systemDns
currently returns.
